renamed shared folder, added shared folder to Dockerfile build steps, testing file resolution ability from build
