+++ challenges.code-workspace
@@ -1,10 +1,9 @@
-
 {
   "folders": [
     {
       "path": ".github",
       "name": "github"
-    },    
+    },
     {
       "path": "src/client",
       "name": "client"
@@ -14,7 +13,7 @@
       "name": "server"
     },
     {
-      "path": "src/CodeWorks_Challenges.shared",
+      "path": "src/shared",
       "name": "shared"
     },
   ],
