+++ src/client/.gitignore
@@ -1,6 +1,7 @@
 .DS_Store
 node_modules
 /dist
+/docs
 
 # local env files
 .env.local
