+++ src/client/src/constants/index.js
@@ -1,4 +1,4 @@
-import * as data from '../../../CodeWorks_Challenges.shared/constants'
+import * as data from '../../../shared/constants'
 
 export const SUBMISSION_TYPES = data.SUBMISSION_TYPES
 export const CATEGORY_TYPES = data.CATEGORY_TYPES
